+++ breeze-core/src/test/kotlin/com/windea/breezeframework/core/tests/Tests.kt
@@ -4,6 +4,7 @@ import com.windea.breezeframework.core.enums.core.LetterCase.*
 import com.windea.breezeframework.core.enums.core.ReferenceCase.*
 import com.windea.breezeframework.core.extensions.*
 import org.junit.*
+import java.time.*
 
 class Tests {
 	@Test
@@ -73,4 +74,12 @@ class Tests {
 	//	println(0.toChar())
 	//	println(25.toChar())
 	//}
+	
+	@Test
+	fun testToString() {
+		println(LocalDate.parse("2014-01-06")) //2014-01-06 OK
+		//PTnHnMnS
+		println(Duration.ofDays(1)) //PT24H
+		println(Duration.ofHours(20)) //PT20H
+	}
 }